refactor: move is_event_loop_running and as_completed to async_utils
This helps separate async features from executor and engine modules, and will prevent circular imports in these modules.
