+++ docs/references/executor.md
@@ -3,4 +3,3 @@
         members:
             - Executor
             - run_async_batch
-            - is_event_loop_running
