+++ ragas/src/ragas/metrics/base.py
@@ -12,9 +12,9 @@ from pydantic import ValidationError
 from tqdm import tqdm
 
 from ragas._analytics import EvaluationEvent, _analytics_batcher
+from ragas.async_utils import is_event_loop_running
 from ragas.callbacks import ChainType, new_group
 from ragas.dataset_schema import MetricAnnotation, MultiTurnSample, SingleTurnSample
-from ragas.executor import is_event_loop_running
 from ragas.losses import BinaryMetricLoss, MSELoss
 from ragas.prompt import FewShotPydanticPrompt, PromptMixin
 from ragas.run_config import RunConfig
