+++ ragas/src/ragas/testset/transforms/engine.py
@@ -6,7 +6,7 @@ import typing as t
 
 from tqdm.auto import tqdm
 
-from ragas.executor import as_completed, is_event_loop_running
+from ragas.async_utils import as_completed, run_async_tasks
 from ragas.run_config import RunConfig
 from ragas.testset.graph import KnowledgeGraph
 from ragas.testset.transforms.base import BaseGraphTransformation
